Cloudinary

+++ app/Http/Controllers/ExperienciaController.php
@@ -120,7 +120,7 @@ class ExperienciaController extends Controller
         $imagePath = null;
         if ($request->hasFile('image')) {
             $optimizer = new ImageOptimizer();
-            $imagePath = '/storage/' . $optimizer->store($request->file('image'));
+            $imagePath = $optimizer->store($request->file('image'));
         }
 
         $post = Post::create([
@@ -283,11 +283,12 @@ class ExperienciaController extends Controller
         $validated = $request->validate($rules);
 
         if ($request->hasFile('image')) {
-            if ($post->image) {
+            // Only delete local images; Cloudinary URLs stay as orphans (cleanup is a separate concern)
+            if ($post->image && str_starts_with($post->image, '/storage/')) {
                 Storage::disk('public')->delete(str_replace('/storage/', '', $post->image));
             }
             $optimizer = new ImageOptimizer();
-            $validated['image'] = '/storage/' . $optimizer->store($request->file('image'));
+            $validated['image'] = $optimizer->store($request->file('image'));
         } else {
             unset($validated['image']);
         }
@@ -309,7 +310,7 @@ class ExperienciaController extends Controller
             abort(403);
         }
 
-        if ($post->image) {
+        if ($post->image && str_starts_with($post->image, '/storage/')) {
             Storage::disk('public')->delete(str_replace('/storage/', '', $post->image));
         }
